feat: add prediction evaluator and blockchain interaction for DePIN plugin
- Implement predictionEvaluator to extract and manage predictions from conversations
- Add PredictionResolver service for handling prediction lifecycle
- Create blockchain helpers for creating and resolving predictions on IoTeX
- Integrate prediction-related template and ABI for smart contract interactions
- Remove deprecated plugin-prediction package and consolidate prediction functionality
