+++ packages/plugin-depin/src/helpers/predictionAbi.ts
