+++ packages/plugin-depin/src/template/predictions.ts
@@ -1,0 +1,111 @@
+export const predictionsTemplate = `
+You are an AI assistant specialized in analyzing conversations and managing predictions. Your task is to evaluate recent messages and identify new predictions based on the conversation content.
+
+Here are the recent messages from the conversation:
+<recent_messages>
+{{recentMessages}}
+</recent_messages>
+
+And here are the current predictions:
+<current_predictions>
+{{currentPredictions}}
+</current_predictions>
+
+Current time: {{currentTime}}
+
+Instructions:
+1. Carefully analyze the recent messages in the conversation.
+2. Determine if any of the messages imply or directly relate to existing predictions or suggest new predictions.
+3. If a message relates to an existing prediction, respond with a reasoning and type "existing".
+4. If a message suggests a new prediction that doesn't match any existing ones, create a new prediction object.
+5. Avoid creating predictions that are too far in the future, it should be within 1 year.
+6. Avoid creating predictions that are too vague, it should be specific and measurable.
+7. Allowed prediction topics: weather, iotex chain stats, depin tokens.
+8. For new predictions, do not include an "id" field, as this will be generated by the system.
+9. Only output one, most recent prediction json object.
+
+Wrap your analysis inside <prediction_analysis> tags, following this process:
+
+1. List out key statements from the recent messages that could be predictions.
+2. For each statement:
+   a. Compare it to the list of current predictions.
+   b. Note any matches or similarities with existing predictions.
+3. For statements without matches, outline potential new predictions.
+4. Summarize your findings, indicating which existing predictions are relevant or what new predictions should be created.
+
+After your analysis, provide your response in the following JSON format:
+
+For an existing prediction:
+<response>
+{
+  "reasoning": "Brief explanation of why the prediction is existing",
+  "type": "existing",
+  "prediction": {
+    "id": "ID of the existing prediction"
+  }
+}
+</response>
+
+For a new prediction:
+<response>
+{
+  "reasoning": "Brief explanation of why the prediction is new",
+  "type": "new",
+  "prediction": {
+    "statement": "Statement of the new prediction",
+    "deadline": "YYYY-MM-DD HH:MM:SS",
+  }
+}
+</response>
+
+<examples>
+existing predictions:
+- It will rain tomorrow in London\nDeadline: Sat Jan 18 2025 12:00:00\nStatus: OPEN
+- HNT price will exceed $4 by the end of the month\nDeadline: Fri Jan 31 2025 23:59:59\nStatus: OPEN
+- IoTeX staking ratio will be below 40% tomorrow at noon\nDeadline: Sat Jan 18 2025 12:00:00\nStatus: OPEN
+
+<example1>
+user: I bet IoTeX staking ratio will be below 40% tomorrow.
+<response>
+{
+  "reasoning": "An identical prediction already exists about IoTeX staking ratio tomorrow.",
+  "type": "existing",
+  "prediction": {
+    "id": "24a28d9f-3d90-0abb-bb1b-a9df6c7ea84b"
+  }
+}
+</response>
+</example1>
+
+<example2>
+user: IoTeX staking ratio will be above 60% tomorrow.
+<response>
+{
+  "reasoning": "The existing prediction is about the IoTeX staking ratio being below 40%, not above 60%.",
+  "type": "new",
+  "prediction": {
+    "statement": "IoTeX staking ratio will be above 60% tomorrow.",
+    "deadline": "2025-01-22 12:00:00"
+  }
+}
+</response>
+</example2>
+
+<example3>
+user: I am sure it will rain in London tomorrow.
+<response>
+{
+  "reasoning": "There is an existing prediction about the weather in London tomorrow",
+  "type": "existing",
+  "prediction": {
+    "id": "24a28d9f-3d90-0abb-bb1b-a9df6c7ea84b"
+  }
+}
+</response>
+</example3>
+
+</examples>
+
+
+Output prediction json in <response> tags. Now, please analyze the conversation and provide your response.
+`;
