--- packages/plugin-prediction/eslint.config.mjs
@@ -1,3 +1,0 @@
-import eslintGlobalConfig from "../../eslint.config.mjs";
-
-export default [...eslintGlobalConfig];
