--- packages/plugin-prediction/package.json
@@ -1,22 +1,0 @@
-{
-    "name": "@elizaos/plugin-prediction",
-    "version": "0.1.6-alpha.4",
-    "main": "dist/index.js",
-    "type": "module",
-    "types": "dist/index.d.ts",
-    "dependencies": {
-      "@elizaos/core": "workspace:*",
-      "tsup": "8.3.5",
-      "axios": "^1.7.9",
-      "@elizaos/plugin-evm": "workspace:*"
-    },
-    "scripts": {
-      "build": "tsup --format esm --dts",
-      "dev": "tsup --format esm --dts --watch",
-      "test": "vitest run",
-      "lint": "eslint --fix  --cache ."
-    },
-    "peerDependencies": {
-      "whatwg-url": "7.1.0"
-    }
-  }
