--- packages/plugin-prediction/tsconfig.json
@@ -1,15 +1,0 @@
-{
-    "extends": "../core/tsconfig.json",
-    "compilerOptions": {
-        "outDir": "dist",
-        "rootDir": "./src",
-        "typeRoots": [
-            "./node_modules/@types",
-            "./src/types"
-        ],
-        "declaration": true
-    },
-    "include": [
-        "src"
-    ]
-}
