--- packages/plugin-prediction/tsup.config.ts
@@ -1,20 +1,0 @@
-import { defineConfig } from "tsup";
-
-export default defineConfig({
-    entry: ["src/index.ts"],
-    outDir: "dist",
-    sourcemap: true,
-    clean: true,
-    format: ["esm"],
-    external: [
-        "dotenv",
-        "fs",
-        "path",
-        "@reflink/reflink",
-        "@node-llama-cpp",
-        "https",
-        "http",
-        "agentkeepalive",
-        "events",
-    ],
-});
